🚚 [RUMF-902] move recorder helpers from E2E to Unit
Because the unit tests have no concept of "EventRegistry", we pass the
segment directly instead of passing the registry.
